Implement seperate uid to App

+++ lib/repositories/items_repository.dart
@@ -1,10 +1,16 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:food/models/item_model.dart';
-
 
 class ItemsRepository {
   Stream<List<ItemModel>> getItemsStream() {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Nie jesteś zalogowany');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('reustarants')
         .orderBy('rating', descending: true)
         .snapshots()
@@ -22,14 +28,20 @@ class ItemsRepository {
     });
   }
 
-  
-
   Future<void> add({
     required String reustarantName,
     required String adresName,
     required String rating,
   }) {
-    return FirebaseFirestore.instance.collection('reustarants').add(
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Nie jesteś zalogowany');
+    }
+    return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
+        .collection('reustarants')
+        .add(
       {
         'name': reustarantName,
         'adres': adresName,
@@ -39,7 +51,13 @@ class ItemsRepository {
   }
 
   Future<void> delete({required String id}) {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('Nie jesteś zalogowany');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('reustarants')
         .doc(id)
         .delete();
